Moves constants dir to client dir for Flow
