+++ client/app/queue/components/IssueListItem.jsx
@@ -12,8 +12,8 @@ import {
   boldText,
   CASE_DISPOSITION_ID_BY_DESCRIPTION
 } from '../constants';
-import ISSUE_INFO from '../../../../constants/ISSUE_INFO.json';
-import VACOLS_DISPOSITIONS_BY_ID from '../../../../constants/VACOLS_DISPOSITIONS_BY_ID.json';
+import ISSUE_INFO from '../../../constants/ISSUE_INFO.json';
+import VACOLS_DISPOSITIONS_BY_ID from '../../../constants/VACOLS_DISPOSITIONS_BY_ID.json';
 
 const minimalLeftPadding = css({ paddingLeft: '0.5rem' });
 const noteMarginTop = css({ marginTop: '1.5rem' });
